requirements.txt
selenium==3.141.0
pytest==6.0.1
allure_python_commons==2.8.13
PyMySQL==0.10.0
Faker==4.1.1
PyYAML==5.3.1
